backend/requirements/common.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
django==3.1.5
django-allauth==0.44.0
django-cors-headers==3.6.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
django-environ
idna==2.10
oauthlib==3.1.0
python3-openid==3.2.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
python-docx

## Auth0
cryptography==3.3.1
drf-jwt==1.17.3
PyJWT
